Added Macro record-enable to perform monitor recording of the calls. macro determines if the extension is setup to record the inbound call and if it is enabled, then the call is saved to the /var/spool/asterisk/monitor directory
Likewise if the call is outbound the same look up is deteremined and if the extension is setup to record outbound calling then the file is saved to the same directory.

Finally, if the call is from a ring group the call is also recorded and saved.

The file name is:
Outbound calls:
CALLFILENAME=OUT-${DATETIME}-${UNIQUEID})
RingGroup calls:
CALLFILENAME=g<ringgroup>-${DATETIME}-${UNIQUEID})
Inbound extension calls
Outbound calls:
CALLFILENAME=${DATETIME}-${UNIQUEID})


git-svn-id: http://svn.freepbx.org/trunk/AMP@366 05cd7311-dd0d-0410-8c4e-c7a75d82e046