DEVICE/USERS separation.  Current cvs users should --force-version 1.10.008. This work removes AMP's extension admin, replacing it with an admin for devices, and an admin to users.  Devices are mapped to users and users to devices.  A user can "logon" to a device (*11) or logoff (*12).
Most of the work is included in this commit.  Very little testing of features has been done, THERE WILL BE BUGS!!.  No testing has been done on voicemail failover or config.  Recording features are not yet functional.

The included 1.10.009beta1 upgrade script will create a device and a user for each existing extension.  The device will be mapped to the user.  Please try it out on a development box and help correct the issues that surface.


git-svn-id: http://svn.freepbx.org/trunk/AMP@540 05cd7311-dd0d-0410-8c4e-c7a75d82e046