initial support for sqlite.
with this patch, freepbx will be able to store it's internal data on 
an sqlite2 database. (I added also an if for pgsql, not tested but should work).
tested on Debian Sarge/Rapid 1.1, php4+sqlite2. if it works on other systems,
feedback would be cool :)

the patch is quite rough, but it seems to work on some places, but still some pages
are not working when using sqlite as a backend. you are not affected if you use mysql, 
and the default settings will continue to be mysql. 

see the sqlite.readme file for more information.

- added a new sql file for sqlite
- added 2 new keys to /etc/amportal.conf:

AMPDENGINE (if not pressent set to mysql)
accepted values: mysql, sqlite, pgsql

AMPDBFILE (used only when AMPDBENGINE is sqlite)
the location of the sqlite2 database.





git-svn-id: http://svn.freepbx.org/trunk/freepbx@1004 05cd7311-dd0d-0410-8c4e-c7a75d82e046