Fix for extension zero issue.
retrieve scripts were using id=0 to be a 'common' set of config options, i guess the idea being allowing some sort of standard set of options for every extensions.  This was disallowing extension zero to be used.  I've changed this to look for -1 instead.  I can't find any current code that was actually storing an id of zero, but obviously there might be some somewhere!  If so, will need chaning to -1 and an upgrade script of some sort written


git-svn-id: http://svn.freepbx.org/freepbx/trunk@1712 05cd7311-dd0d-0410-8c4e-c7a75d82e046